// polaris/src/components/common/models/tool.ts
export interface PSToolModel {
    uid: string
    title: string
    name: string
    keywords: string
    description: string
    status: number
    cover: string
    owner: string
    discover: number
    version: string
    url: string
    lang: string
    create_time: string
    update_time: string
}